pyproject.toml
[tool.poetry]
name = "orango"
version = "0.1.0"
description = "Secure sandboxed cloud and edge environments made for AI agents and AI apps"
authors = ["Orango AI <hello@orango.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
langchain = "^0.2.12"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
requests = "^2.32.3"
requests-mock = "^1.12.1"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
